requirements.txt
django==1.8.2
django-import-export
